--- app/api/bgm_tv_auto_tracker/auth/scheme.py
@@ -1,3 +1,0 @@
-from fastapi.security import APIKeyCookie
-
-cookie_scheme = APIKeyCookie(name='bgm-tv-auto-tracker')
